Merge branch 'sfc-next'
Shradha Shah says:

====================
sfc: ndo_get_phys_port_id, vadaptor stats and PF unload when Vf's assigned to guest

This is the third and last instalment of SRIOV for EF10 patches.

This patch set includes implementation of ndo_get_phys_port_id
and changes to the MAC statistics code in order to support
vadaptor statistics.

It also includes code to deal with PF unload when Vf's are still
assigned to the guest.

The first couple of patches create sysfs files for physical port
and link control flags which are particularly useful when we have
enabled a large number of VF's.

These patches have been tested with and without CONFIG_SFC_SRIOV.
The creation and content of the sysfs files has been tested.
The statistics are tested using ethtool for monitoring.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: df905ceae3883f024282d00824ce33040097ac86 ]