Merge branch 'switchdev-next'
Scott Feldman says:

====================
switchdev: documentation updates

Update the switchdev documentation to clarify that for IPV4 FIB object, the add
op is used for both adding and modifying the FIB entry on the device.

I found a few other white-space and grammer issues to fix, and to clarify the
static FDB ops.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 37c8e2b069c7ccff3ff6a356f82ed11abab5fb99 ]